+++ examples/express.js
@@ -1,6 +1,6 @@
 import express from 'express';
 
-import { errors, status } from '../src';
+import { createError, errors, status } from '../src';
 
 const app = express();
 
@@ -10,7 +10,8 @@ app.get('/api', (req, res) => {
     });
 });
 
-app.get('/404', (req, res, next) => next(new errors.NotFoundError('Nothing to see here!')));
-app.get('/500', (req, res, next) => next(new errors.InternalServerError('Something broke!')));
+app.get('/errors/404', (req, res, next) => next(new errors.NotFoundError('Nothing to see here!')));
+app.get('/errors/500', (req, res, next) => next(new errors.InternalServerError('Something broke!')));
+app.get('/custom/:code', (req, res, next) => next(createError(req.params.code)));
 
 app.listen(3000);