+++ examples/koa.js
@@ -1,7 +1,7 @@
 import Koa from 'koa';
 import Router from 'koa-router';
 
-import { errors, status } from '../src';
+import { createError, errors, status } from '../src';
 
 const app = new Koa();
 const router = new Router();
@@ -16,14 +16,20 @@ router.get('/api', async (ctx, next) => {
     };
 });
 
-router.get('/404', () => {
+router.get('/errors/404', () => {
     throw new errors.NotFoundError('Nothing to see here!'); // error exposed to client by default
 });
 
-router.get('/500', () => {
+router.get('/errors/500', () => {
     throw new errors.InternalServerError('Something broke!'); // logs actual error message but only shows Internal Server Error publicly
+});
+
+router.get('/custom/:code', (ctx) => {
+    throw createError(ctx.params.code); // throws the error that corresponds to the status code provided
 });
 
 app.use(router.routes());
 
-app.listen(3000);
+app.listen(3000, () => {
+    console.log('App is listening on port 3000...');
+});