+++ test/errors/index.js
@@ -7,7 +7,7 @@
 
 import test from 'ava';
 
-import { errors, status } from '../../src/index';
+import { createError, errors, status } from '../../src/index';
 
 test('Error classes', (t) => {
     let err;
@@ -17,11 +17,13 @@ test('Error classes', (t) => {
             err = new errors[e]();
 
             t.true(err.expose);
+            t.is(err.status, errors[e].code());
             t.is(err.message, status[err.status]);
         } else if (errors[e].prototype instanceof errors.ServerError) {
             err = new errors[e]();
 
             t.false(err.expose);
+            t.is(err.status, errors[e].code());
             t.is(err.message, status[err.status]);
         } else if (errors[e].name === 'ClientError') {
             err = new errors[e](400, 'Message', { props: { custom: 1, expose: 0 } }); // expose should be ignored
@@ -49,4 +51,30 @@ test('Error classes', (t) => {
             t.fail(`Unexpected member of errors module: ${errors[e]}`); // should not reach this
         }
     }
+});
+
+test('Create client error using createError function', (t) => {
+    const notFoundErr = createError(404);
+
+    t.true(notFoundErr.expose);
+    t.is(notFoundErr.status, status.NOT_FOUND);
+    t.is(notFoundErr.message, status[notFoundErr.status]);
+
+    const notFoundErrWithMessage = createError(404, 'Custom message');
+
+    t.is(notFoundErrWithMessage.message, 'Custom message');
+
+    const notFoundErrWithExposeFalse = createError(404, undefined, { expose: false });
+
+    t.is(notFoundErrWithExposeFalse.expose, false);
+
+    const notFoundErrWithCustomProp = createError(404, undefined, { props: { x: 10 } });
+
+    t.is(notFoundErrWithCustomProp.x, 10);
+});
+
+test('Create error using createError function should fail when statusCode is not 4xx or 5xx', (t) => {
+    t.throws(() => {
+        createError(200);
+    }, Error);
 });